+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Always respect automatic filtering of control characters in field values in QueryType\Update\Query\Document
+- Remove the field modifier along with the value(s) and boost in QueryType\Update\Query\Document::removeField()
 
 
 ## [6.1.5]