+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 ### Added
+- PHP 8.1 support
 - QueryType\Update\Query\Document::setFields() to set all fields on a Document
 
 ### Fixed